modules/logger_grpc/requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
kafka-python==2.0.2
protobuf==3.19.4
six==1.16.0
